refactor eviction in filesystem middleware to use shared utils
